+++ packages/ui/react/src/index.ts
@@ -2,6 +2,6 @@
 export * from '@favolink-ui/hooks';
 export * from '@favolink-ui/icons';
 export * from '@favolink-ui/system';
-export * from '@favolink-ui/styles';
+export * from './styles';
 export * from '@favolink-ui/utils';
 export * from './components';
