+++ packages/ui/react/src/styles/utils/flex-shrink.css.ts
