+++ packages/ui/react/src/styles/utils/truncate.css.ts
