--- packages/ui/styles/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@favolink-ui/styles",
-  "version": "0.0.0",
-  "private": true,
-  "main": "src/index.ts",
-  "scripts": {
-    "lint": "eslint \"**/*.{js,jsx,cjs,mjs,ts,tsx,cts,mts}\""
-  },
-  "dependencies": {
-    "@favolink-ui/utils": "workspace:*",
-    "@vanilla-extract/css": "^1.14.1",
-    "@vanilla-extract/dynamic": "^2.1.0",
-    "react": "^18.2.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.3.3"
-  }
-}
